refactor(framework): Move object_ref
