+++ framework/py/flwr/server/run_serverapp.py
@@ -19,9 +19,9 @@ from logging import DEBUG
 
 from flwr.app import Context
 from flwr.common.logger import log
-from flwr.common.object_ref import load_app
 from flwr.serverapp.grid import Grid
 from flwr.serverapp.server_app import LoadServerAppError, ServerApp
+from flwr.supercore.object_ref import load_app
 
 
 def run(
