+++ framework/py/flwr/supercore/task_process/agent/run_agentapp.py
@@ -35,7 +35,6 @@ from flwr.common.config import (
 from flwr.common.constant import RUNTIME_DEPENDENCY_INSTALL, SubStatus
 from flwr.common.exit import ExitCode, flwr_exit, register_signal_handlers
 from flwr.common.logger import flush_logs, log, start_log_uploader, stop_log_uploader
-from flwr.common.object_ref import load_app
 from flwr.common.serde import (
     context_from_proto,
     context_to_proto,
@@ -49,6 +48,7 @@ from flwr.proto.appio_pb2 import (  # pylint: disable=E0611
 )
 from flwr.supercore.app_utils import start_parent_process_monitor
 from flwr.supercore.heartbeat import HeartbeatSender, make_task_heartbeat_fn_grpc
+from flwr.supercore.object_ref import load_app
 from flwr.supercore.superexec.dependency_installer import (
     RuntimeDependencyInstallationError,
     cleanup_app_runtime_environment,
